Split test/acoustid/test_acoustid.py by module
The file covered two modules, so rename it to one and extract the other:

- test/acoustid/test_acoustid.py -> test_json_helpers.py (parse_recording,
  recording_has_metadata from picard.acoustid.json_helpers)
- new test/acoustid/test_recordings.py (Recording, max_source_count,
  parse_recording_map from picard.acoustid.recordings)

The shared AcoustIDTest fixture stays with the json_helpers tests that use
it; the recordings tests already subclass PicardTestCase directly. No test
logic changed; the method count is unchanged (7 = 4 + 3).
