+++ test/acoustid/test_json_helpers.py
@@ -31,11 +31,6 @@ from picard.acoustid.json_helpers import (
     parse_recording,
     recording_has_metadata,
 )
-from picard.acoustid.recordings import (
-    Recording,
-    max_source_count,
-    parse_recording_map,
-)
 from picard.mbjson import recording_to_metadata
 from picard.metadata import Metadata
 from picard.track import Track
@@ -122,56 +117,6 @@ class NullRecordingTest(AcoustIDTest):
         parsed_recording = parse_recording(self.json_doc)
         recording_to_metadata(parsed_recording, m, t)
         self.assertEqual(m, {})
-
-
-class MaxSourceCountTest(PicardTestCase):
-    def test_max_source_count(self):
-        recordings = (
-            Recording({}, sources=5),
-            Recording({}, sources=13),
-            Recording({}, sources=12),
-        )
-        self.assertEqual(13, max_source_count(recordings))
-
-    def test_max_source_count_no_recordings(self):
-        self.assertEqual(1, max_source_count([]))
-
-    def test_max_source_count_smaller_1(self):
-        recordings = (Recording({}, sources=0),)
-        self.assertEqual(1, max_source_count(recordings))
-
-
-class ParseRecordingMapTest(PicardTestCase):
-    def test_parse_recording_map(self):
-        recordings = {
-            "cb127606-8e3d-4dcf-9902-69c7a9b59bc9": {
-                "d12e0535-3b2f-4987-8b03-63d85ef57fdd": Recording(
-                    {
-                        "id": "d12e0535-3b2f-4987-8b03-63d85ef57fdd",
-                    }
-                ),
-                "e8a313ee-b723-4b48-9395-6c8e026fc9e0": Recording(
-                    {"id": "e8a313ee-b723-4b48-9395-6c8e026fc9e0"},
-                    sources=4,
-                    result_score=0.5,
-                ),
-            }
-        }
-        expected = [
-            {
-                "id": "d12e0535-3b2f-4987-8b03-63d85ef57fdd",
-                "acoustid": "cb127606-8e3d-4dcf-9902-69c7a9b59bc9",
-                "score": 25,
-                "sources": 1,
-            },
-            {
-                "id": "e8a313ee-b723-4b48-9395-6c8e026fc9e0",
-                "acoustid": "cb127606-8e3d-4dcf-9902-69c7a9b59bc9",
-                "score": 50,
-                "sources": 4,
-            },
-        ]
-        self.assertEqual(expected, list(parse_recording_map(recordings)))
 
 
 class RecordingHasMetadataTest(PicardTestCase):
